+++ CHANGELOG.md
@@ -12,6 +12,8 @@ port A.
   - Faster and more deterministic sampling
   - Implement customizable averaging windows based on crank angles
 - Fix occasional miss due to incorrectly descheduling events
+- Fix bug with `get bootloader`
+- Move task execution to systick timer
 
 ### 1.1.0 (2019 Dec 14)
 Introduces a breaking change with the console format for tables, as well as